add a protocol for externally controlled display brightness
For some use cases, like HDR content, content-adaptive brightness or automatic
brightness control, the compositor has to control the brightness of backlight
devices. However, these interfaces are sometimes unsable or require root
privileges, so implementing support for them directly in the compositor isn't
ideal.
This protocol allows an external process (powerdevil) to advertise brightness
devices for KWin to control, which circumvents those problems and allows for
code reuse with the existing implementation for X11